+++ tests/cli/test_performance_cli.py
@@ -1,6 +1,6 @@
 import os
 
-from tests.helpers.helpers import (
+from tests.helpers import (
     assert_empty_df,
     assert_empty_list,
     pandas_error,
