+++ tests/cli/test_stats_cli.py
@@ -2,7 +2,7 @@ import os
 
 import pytest
 
-from tests.helpers.helpers import (
+from tests.helpers import (
     assert_empty_df,
     load_pandas,
     pandas_error,
