+++ tests/cli/test_tolerance_cli.py
@@ -1,6 +1,6 @@
 import os
 
-from tests.helpers.helpers import (
+from tests.helpers import (
     assert_empty_df,
     load_pandas,
     pandas_error,
